Extract out new module `TokenName.Gen`.
