+++ lib/primitive/lib/Cardano/Wallet/Primitive/Types/TokenPolicy/Gen.hs
@@ -1,19 +1,14 @@
 module Cardano.Wallet.Primitive.Types.TokenPolicy.Gen
     (
     -- * Generators and shrinkers
-      genTokenName
-    , genTokenNameLargeRange
-    , genTokenPolicyId
+      genTokenPolicyId
     , genTokenPolicyIdLargeRange
-    , shrinkTokenName
     , shrinkTokenPolicyId
 
     -- * Test values
-    , testTokenNames
     , testTokenPolicyIds
 
     -- * Creation of test values
-    , mkTokenName
     , mkTokenPolicyId
 
     ) where
@@ -22,9 +17,6 @@ import Prelude
 
 import Cardano.Wallet.Primitive.Types.Hash
     ( Hash (..)
-    )
-import Cardano.Wallet.Primitive.Types.TokenName
-    ( TokenName (..)
     )
 import Cardano.Wallet.Primitive.Types.TokenPolicy
     ( TokenPolicyId (..)
@@ -43,29 +35,7 @@ import Test.QuickCheck
     )
 
 import qualified Data.ByteString as BS
-import qualified Data.ByteString.Char8 as B8
 import qualified Data.Text as T
-
---------------------------------------------------------------------------------
--- Token names chosen from a range that depends on the size parameter
---------------------------------------------------------------------------------
-
-genTokenName :: Gen TokenName
-genTokenName = sized $ \n -> elements $ take (max 1 n) testTokenNames
-
-shrinkTokenName :: TokenName -> [TokenName]
-shrinkTokenName i
-    | i == simplest = []
-    | otherwise = [simplest]
-  where
-    simplest = head testTokenNames
-
---------------------------------------------------------------------------------
--- Token names chosen from a large range (to minimize the risk of collisions)
---------------------------------------------------------------------------------
-
-genTokenNameLargeRange :: Gen TokenName
-genTokenNameLargeRange = UnsafeTokenName . BS.pack <$> vector 32
 
 --------------------------------------------------------------------------------
 -- Token policy identifiers chosen from a range that depends on the size
@@ -94,14 +64,8 @@ genTokenPolicyIdLargeRange = UnsafeTokenPolicyId . Hash . BS.pack <$> vector 28
 -- Internal utilities
 --------------------------------------------------------------------------------
 
-testTokenNames :: [TokenName]
-testTokenNames = mkTokenName <$> ['A' .. 'Z']
-
 testTokenPolicyIds :: [TokenPolicyId]
 testTokenPolicyIds = mkTokenPolicyId <$> mkTokenPolicyIdValidChars
-
-mkTokenName :: Char -> TokenName
-mkTokenName = UnsafeTokenName . B8.snoc "Token"
 
 -- The set of characters that can be passed to the 'mkTokenPolicyId' function.
 --
